ai-hedge-fund-main/requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
langchain-core>=0.1.0
langgraph>=0.0.11
python-dotenv>=1.0.0
yfinance>=0.2.36 